Include response chunking info in nonce (#3)
If a given delta response contains too many resources, the server will
break it up into multiple responses. However, this means the client does
not know whether it received all the resources for its subscription.
This is especially relevant for wildcard subscriptions, for which the
client does not know the resources ahead of time and therefore cannot
wait for them explicitly. By returning additional metadata in the nonce
(there is no field for this in the delta discovery response, though I'm
hoping that will change https://github.com/cncf/xds/issues/99), the
client can know if the server chunked the response, and react
accordingly.