Pool chunk slices instead of reusing them
Similar to #2, where buffers are reused instead of pooled. Reusing a buffer
instead of pooling it means its capacity is never released. Therefore if a large
spike in traffic causes that buffer to grow, the resources it holds onto are
never returned even as the traffic decreases. This causes those large buffers to
stick around and not get used.

This change uses a `sync.Pool` instead of each `deltaSender` having its own
buffer. This way buffers are automatically released when they aren't used.