Remove explicit handling of unsupported resource types
Because there is no explicit provision for unknown resource types in the xDS
protocol, the server implementation had to make a bunch of assumptions about
how it should respond. These assumptions incidentally violated the xDS protocol
NACK/ACK flow. Instead, the behavior should be handed off to the
`ResourceLocator` implementation so that it can decide what to do. This also
allows the simplification of the `ResourceLocator` interface, which is an added
benefit.

Note that this change is backwards incompatible, so will need to be bumped
carefully.