Implement ADS client
This change includes the new `diderot.ADSClient` type, which is a generic ADS
client that can be used to talk to any ADS server (including notably the
`diderot.ADSServer` :)). It also includes a number of improvements to the way
glob collection URLs are handled, and makes it easier to manage them.