Change API for checking resource queue time
This relatively small change notifies the stats handler of the resources being
sent by the server right as they are about to be sent, rather than every time
they are queued. This allows the stats handler to compute things in bulk rather
than one at a time, allowing room for improvement.