Do not print large resources
This was a very stupid log message. If the resource is larger than 4MB it will
quickly flood the logs. The resource itself is preserved in the
`serverstats.ResourceOverMaxSize` message, if the stats handler wants to log it
or inspect it if it wants, but `diderot` itself should not log it.