Wire up sent nonces
By adding new, more granular server event types, event handlers now have much
more visibility into what the server is doing. Specifically, it can now track
pending nonces and in-flight requests. Similarly, because the `RequestReceived`
fires when the request is received and `RequestProcessed` fires when the request
is *processed*, the two events can be used to track whether the ResourceLocator
is hanging.